Separate index into multiple files
